first pass refactoring to split cli and library
